fix(android): сборка (getLocationOnScreen int[], runOnUiThread) + переходы в Медиатеку по вкладкам
- Sheets.insideSheet: корректный int[] для getLocationOnScreen
- MainActivity: слушатель плеера через runOnUiThread
- showLibraryTab(tab): «Офлайн» → Скачано, «Недавние» → История (как /library?tab=...)

+++ android/app/src/main/java/com/anibeat/app/ui/Sheets.java
@@ -106,7 +106,7 @@ public class Sheets extends FrameLayout {
 
     /** true, если касание попало в сам лист шторки, а не в затемнение. */
     private boolean insideSheet(MotionEvent e) {
-        float[] loc = new float[2];
+        int[] loc = new int[2];
         sheet.getLocationOnScreen(loc);
         return e.getRawY() >= loc[1];
     }
